Refactor Cast (#26550)
* Refactor Cast

* Fix tests & address comments

* Update reqs
