+++ script/gen_requirements_all.py
@@ -10,8 +10,8 @@ import sys
 from script.hassfest.model import Integration
 
 COMMENT_REQUIREMENTS = (
-    "Adafruit-DHT",
     "Adafruit_BBIO",
+    "Adafruit-DHT",
     "avion",
     "beacontools",
     "blinkt",
@@ -26,7 +26,6 @@ COMMENT_REQUIREMENTS = (
     "i2csense",
     "opencv-python-headless",
     "py_noaa",
-    "VL53L1X2",
     "pybluez",
     "pycups",
     "PySwitchbot",
@@ -39,11 +38,11 @@ COMMENT_REQUIREMENTS = (
     "RPi.GPIO",
     "smbus-cffi",
     "tensorflow",
+    "VL53L1X2",
 )
 
 TEST_REQUIREMENTS = (
     "adguardhome",
-    "ambiclimate",
     "aio_geojson_geonetnz_quakes",
     "aioambient",
     "aioautomatic",
@@ -52,14 +51,16 @@ TEST_REQUIREMENTS = (
     "aiohttp_cors",
     "aiohue",
     "aionotion",
-    "aiounifi",
     "aioswitcher",
+    "aiounifi",
     "aiowwlln",
+    "ambiclimate",
     "androidtv",
     "apns2",
     "aprslib",
     "av",
     "axis",
+    "bellows-homeassistant",
     "caldav",
     "coinmarketcap",
     "defusedxml",
@@ -99,7 +100,6 @@ TEST_REQUIREMENTS = (
     "libpurecool",
     "libsoundtouch",
     "luftdaten",
-    "pyMetno",
     "mbddns",
     "mficlient",
     "minio",
@@ -115,44 +115,49 @@ TEST_REQUIREMENTS = (
     "ptvsd",
     "pushbullet.py",
     "py-canary",
+    "py17track",
     "pyblackbird",
+    "pychromecast",
     "pydeconz",
     "pydispatcher",
     "pyheos",
     "pyhomematic",
+    "pyHS100",
     "pyiqvia",
     "pylinky",
     "pylitejet",
+    "pyMetno",
     "pymfy",
     "pymonoprice",
+    "PyNaCl",
     "pynws",
     "pynx584",
     "pyopenuv",
     "pyotp",
     "pyps4-homeassistant",
+    "pyqwikswitch",
+    "PyRMVtransport",
     "pysma",
     "pysmartapp",
     "pysmartthings",
     "pysonos",
-    "pyqwikswitch",
-    "PyRMVtransport",
-    "PyTransportNSW",
     "pyspcwebgw",
+    "python_awair",
     "python-forecastio",
     "python-nest",
-    "python_awair",
     "python-velbus",
+    "pythonwhois",
     "pytradfri[async]",
+    "PyTransportNSW",
     "pyunifi",
     "pyupnp-async",
     "pyvesync",
     "pywebpush",
-    "pyHS100",
-    "PyNaCl",
     "regenmaschine",
     "restrictedpython",
     "rflink",
     "ring_doorbell",
+    "ruamel.yaml",
     "rxv",
     "simplisafe-python",
     "sleepyq",
@@ -166,16 +171,12 @@ TEST_REQUIREMENTS = (
     "twentemilieu",
     "uvcclient",
     "vsure",
-    "warrant",
-    "pythonwhois",
-    "wakeonlan",
     "vultr",
+    "wakeonlan",
+    "warrant",
     "YesssSMS",
-    "ruamel.yaml",
     "zeroconf",
     "zigpy-homeassistant",
-    "bellows-homeassistant",
-    "py17track",
 )
 
 IGNORE_PIN = ("colorlog>2.1,<3", "keyring>=9.3,<10.0", "urllib3")
